+++ android/src/main/java/com/twiliorn/library/CustomTwilioVideoView.java
@@ -153,6 +153,8 @@ public class CustomTwilioVideoView extends View
     private static final String TRACK_NAME_MICROPHONE = "microphone";
     private static final String TRACK_NAME_SCREEN = "screen";
     private static final String TRACK_NAME_DATA = "data";
+    private static final String PRODUCT_NAME_KEY = "com.twilio.video.product.name";
+    private static final String PRODUCT_VERSION_KEY = "com.twilio.video.product.version";
     private static final int REQUEST_MEDIA_PROJECTION = 100;
     private boolean enableRemoteAudio = false;
     private boolean enableNetworkQualityReporting = false;
@@ -344,6 +346,10 @@ public class CustomTwilioVideoView extends View
         super(context);
         this.themedReactContext = context;
         this.eventEmitter = themedReactContext.getJSModule(RCTEventEmitter.class);
+
+        // Set properties for Video Insights reporting 
+        System.setProperty(PRODUCT_NAME_KEY, TwilioVideoConstants.kTwilioVideoReactNativeName);
+        System.setProperty(PRODUCT_VERSION_KEY, TwilioVideoConstants.kTwilioVideoReactNativeVersion);
 
         // add lifecycle for onResume and on onPause
         themedReactContext.addLifecycleEventListener(this);